Script/requirements.txt
requests
pyyaml
sqlite3

